requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
